# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Employee:
    properties:
      department_id:
        type: integer
      division_id:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        description: db:"id" указывает на столбец id в базе данных
        type: integer
      last_name:
        type: string
      manager_id:
        type: integer
      phone:
        type: string
      photo_path:
        description: Make the first letter uppercase
        type: string
      position:
        type: string
      role:
        type: string
      unit_id:
        type: integer
    type: object
  models.Filters:
    properties:
      key:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://example.com/contact
  description: API для управления сотрудниками и их иерархией.
  termsOfService: http://example.com/terms/
  title: Employee Management API
  version: "1.0"
paths:
  /employees:
    get:
      description: Возвращает список всех сотрудников с возможностью фильтрации
      parameters:
      - description: Имя
        in: query
        name: first_name
        type: string
      - description: Фамилия
        in: query
        name: last_name
        type: string
      - description: Должность
        in: query
        name: position
        type: string
      - description: Роль
        in: query
        name: role
        type: string
      - description: Телефон
        in: query
        name: phone
        type: string
      - description: Почта
        in: query
        name: email
        type: string
      - description: ID руководителя
        in: query
        name: manager_id
        type: integer
      - description: ID департамента
        in: query
        name: department_id
        type: integer
      - description: ID подразделения
        in: query
        name: division_id
        type: integer
      - description: ID отдела
        in: query
        name: unit_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получение всех сотрудников
      tags:
      - Employees
  /employees/hierarchy:
    get:
      description: Возвращает иерархию сотрудников по manager_id или department_id
      parameters:
      - description: ID руководителя
        in: query
        name: manager_id
        type: integer
      - description: ID отдела
        in: query
        name: department_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получение иерархии сотрудников
      tags:
      - Employees
  /filters:
    get:
      description: Возвращает список фильтров для сотрудников
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Filters'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получение списка доступных фильтров
      tags:
      - Employees
  /section_name:
    get:
      description: Возвращает название отдела организации по его ID
      parameters:
      - description: ID отдела
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получение названия отдела по ID
      tags:
      - Sections
swagger: "2.0"